Save original image URL, then restore on error
When you save the page with Ctrl+S, Firefox and Chrome download all of
the images and save them in a separate _files directory next to the HTML
file. If you reupload the HTML file somewhere else without the _files
directory, all of the images are now gone!

This commit saves the original image URL in the 'data-original-src'
attribute. Then, if the image fails to load, we reset the real 'src' to
the saved image URL and try again.